[Fix] Fix lifo testing. Assert exception called
